vkDataCollection/requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
idna==2.10
multidict==5.0.0
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.6.2
